+++ R/get_badges.R
@@ -1,41 +1,78 @@
-#' Return bagdes for a given workshop.
+#' Return badges for a given workshop.
 #'
 #' @param workshop_id workshop identifier, see [workshops_df()].
-#' @param lang the language for the badges.
+#' @param lang the language for the badges (either "en" or "fr").
+#' @param clip a logical. Should the badges be copied to the clipboard.
 #' @export
 #' @examples
-#' get_badges(1)
-#' get_badges(4, lang = "fr")
+#' get_badges(1, clip = FALSE)
+#' get_badges(4, lang = "fr", clip = FALSE)
 
-get_badges <- function(workshop_id , lang=c("en", "fr")) {
+get_badges <- function(workshop_id, lang = c("en", "fr"), clip = TRUE) {
 
   lang <- match.arg(lang)
   id2 <- sprintf("%02d", workshop_id)
 
   if (lang == "en") {
-    prs <- "Presentation"
+    prs <- "Slides"
     atl <- "Workshop"
   } else {
-    prs <- "Pr\u00e9sentation"
+    prs <- "Diapos"
     atl <- "Atelier"
   }
 
-  tra <- paste0("[![Build Status](https://img.shields.io/travis/QCBSRworkshops/workshop",
-  id2,
-  "/dev?style=for-the-badge)](https://travis-ci.org/QCBSRworkshops/workshop", id2, ")")
+  wiq <- glue("https://wiki.qcbs.ca/r_{tolower(prs)}{workshop_id}")
 
-  pres <- paste0(
-    "[![](https://img.shields.io/static/v1?style=for-the-badge&label=",
-    prs, "&message=", atl,
-    "%20", workshop_id,
-    "&color=blue)](https://qcbsrworkshops.github.io/workshop",
-    id2, "/workshop", id2, "-", lang, "/workshop", id2, "-", lang, ".html)")
+  wid <- glue("workshop{workshop_id}")
+  wid2 <- sprintf("workshop%02d", workshop_id)
+  basurl <- glue("https://qcbsrworkshops.github.io/{wid2}/{wid2}-{lang}")
 
-  wiki <- paste0("[![](https://img.shields.io/static/v1?style=for-the-badge&label=Wiki&message=",
-    atl, "%20", workshop_id, "&color=violet)](https://wiki.qcbs.ca/r_",
-    tolower(atl), workshop_id , ")")
+  tra <- get_badge_travis("QCBSRworkshops", wid2, "dev")
+  gh <- get_badge_indiv("QCBSRworkshops", wid2, "dev", color = "6f42c1")
+  html <- get_badge_one(url = glue("{basurl}/{wid2}-{lang}.html"),
+    logo = "html5", label = prs, message = id2, color="red")
+  pdf <- get_badge_one(url = glue("?"),
+    logo = "adobe-acrobat-reader", label = prs, message = id2, color="red")
+  wiki <- get_badge_one(url = wiq, logo="wikipedia", label = "wiki",
+    message = id2)
+  scr <- get_badge_one(logo = "r", label = "script", message = id2, color = "2a50b8")
 
-  cat(tra, pres, wiki, sep="\n")
+  out <- paste(tra, gh, wiki, html, pdf, scr, sep="\n")
+  cat(out, "\n")
+  if (clip) write_clip(out)
 
-  invisible(paste(tra, pres, wiki, sep="\n"))
+  info_msg("badges copied in your clipboard!")
+  invisible(out)
 }
+
+
+get_badge_one <- function(url = NULL, main = url_shield.io(),
+  style = "flat-square", label=NULL, message=NULL, color = NULL, logo = NULL, logoColor = NULL,
+  alt = "badge") {
+
+  out <- glue("{main}style={style}")
+  if (!is.null(label)) out <- glue("{out}&label={label}")
+  if (!is.null(message)) out <- glue("{out}&message={message}")
+  if (!is.null(color)) out <- glue("{out}&color={color}")
+  if (!is.null(logo)) out <- glue("{out}&logo={logo}")
+  if (!is.null(logoColor)) out <- glue("{out}&logoColor={logoColor}")
+  out <- glue("![{alt}]({out})")
+
+  if (!is.null(url)) {
+    sprintf("[%s](%s)", out, url)
+  } else out
+}
+
+get_badge_travis <- function(user, repo, branch, ...) {
+  main <- glue("https://img.shields.io/travis/{user}/{repo}/{branch}?")
+  url <- glue("https://travis-ci.org/{user}/{repo}")
+  get_badge_one(url, main, logo="travis", ...)
+}
+
+get_badge_indiv <- function(user, repo, branch, ...) {
+  url <- glue("https://github.com/{user}/{repo}")
+  get_badge_one(url, logo="github", label = "repo", message = branch, ...)
+}
+
+url_shield.io <- function() "https://img.shields.io/static/v1?"
+# c("plastic", "flat", "flat-square", "for-the-badge", "social")